# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rewards": {
      "get": {
        "operationId": "RewardsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RewardGameDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "The Rewards screen: every card with its countdown",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/rewards/winners": {
      "get": {
        "operationId": "RewardsController_allWinners",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "One exact draw day.",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          },
          {
            "name": "date_from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RewardWinnerDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Winners across every reward, filterable by date",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/rewards/{slug}": {
      "get": {
        "operationId": "RewardsController_detail",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/RewardDetailDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One reward: pot, countdown, my entries, prizes, winners",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/rewards/{slug}/winners": {
      "get": {
        "operationId": "RewardsController_winners",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "description": "One exact draw day.",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          },
          {
            "name": "date_from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-01",
              "type": "string"
            }
          },
          {
            "name": "date_to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-08-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RewardWinnerDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Winners of one reward, filterable by date",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/rewards/{slug}/entries": {
      "post": {
        "operationId": "RewardsController_buyEntries",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyEntriesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/RewardEntriesDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buys entries into a draw with gems",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/rewards/{slug}/play": {
      "post": {
        "operationId": "RewardsController_play",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/RewardPlayDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pays for one instant play and resolves it",
        "tags": [
          "rewards"
        ]
      }
    },
...
    "/api/support/tickets": {
      "post": {
        "operationId": "SupportController_createTicket",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SupportTicketDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Raise a support ticket, problem report, or feedback",
        "tags": [
          "support"
        ]
      },
      "get": {
        "operationId": "SupportController_listTickets",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Rows per page, max 100.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/SupportTicketDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "My tickets",
        "tags": [
          "support"
        ]
      }
    },
    "/api/support/tickets/{id}": {
      "get": {
        "operationId": "SupportController_getTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/SupportTicketDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get one of my tickets",
        "tags": [
          "support"
        ]
      }
    },
    "/api/support/faqs": {
      "get": {
        "operationId": "SupportController_listFaqs",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category slug from GET /api/support/faq-categories. Omit for every FAQ.",
            "schema": {
              "example": "payment",
              "type": "string"
            }
          },
          {
            "name": "category_id",
            "required": false,
            "in": "query",
            "description": "Same filter by id, for callers that already hold one.",
            "schema": {
              "example": "0f1b2c3d-4e5f-6789-abcd-ef0123456789",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/FaqListDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Every FAQ with its categories, optionally filtered to one category. Public.",
        "tags": [
          "support"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coinzu API",
    "description": "Coinzu backend API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RewardGameDto": {
        "type": "object",
        "properties": {
          "cz_reward_game_id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string",
            "example": "daily_lucky_draw"
          },
          "kind": {
            "type": "string",
            "enum": [
              "instant",
              "draw"
            ]
          },
          "cadence": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly"
            ]
          },
          "title": {
            "type": "string",
            "example": "Daily Lucky Draw"
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true
          },
          "headline_prize_coins": {
            "type": "number",
            "example": 5000
          },
          "entry_cost_gems": {
            "type": "number",
            "example": 10
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "coming_soon",
              "paused"
            ]
          },
          "is_playable": {
            "type": "boolean"
          },
          "ends_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "seconds_remaining": {
            "type": "number",
            "nullable": true,
            "example": 86694
          },
          "my_entries": {
            "type": "number",
            "example": 0
          },
          "cz_lucky_draw_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "display_order": {
            "type": "number",
            "example": 1
          }
        },
        "required": [
          "cz_reward_game_id",
          "slug",
          "kind",
          "cadence",
          "title",
          "subtitle",
          "icon_url",
          "headline_prize_coins",
          "entry_cost_gems",
          "status",
          "is_playable",
          "ends_at",
          "seconds_remaining",
          "my_entries",
          "cz_lucky_draw_id",
          "display_order"
        ]
      },
      "RewardWinnerDto": {
        "type": "object",
        "properties": {
          "cz_lucky_draw_winner_id": {
            "type": "string",
            "format": "uuid"
          },
          "cz_lucky_draw_id": {
            "type": "string",
            "format": "uuid"
          },
          "game_slug": {
            "type": "string",
            "nullable": true
          },
          "game_title": {
            "type": "string",
            "nullable": true
          },
          "period_key": {
            "type": "string",
            "nullable": true
          },
          "draw_date": {
            "type": "string",
            "format": "date-time"
          },
          "rank": {
            "type": "number",
            "example": 1
          },
          "masked_name": {
            "type": "string",
            "example": "prash@****.com"
          },
          "avatar_url": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "prize_coins": {
            "type": "number"
          },
          "prize_gems": {
            "type": "number"
          },
          "entries_held": {
            "type": "number"
          }
        },
        "required": [
          "cz_lucky_draw_winner_id",
          "cz_lucky_draw_id",
          "game_slug",
          "game_title",
          "period_key",
          "draw_date",
          "rank",
          "masked_name",
          "avatar_url",
          "country",
          "prize_coins",
          "prize_gems",
          "entries_held"
        ]
      },
      "RewardStepDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "title",
          "description",
          "icon_url"
        ]
      },
      "RewardPrizeDto": {
        "type": "object",
        "properties": {
          "cz_reward_prize_id": {
            "type": "string",
            "format": "uuid"
          },
          "rank": {
            "type": "number",
            "example": 1
          },
          "label": {
            "type": "string",
            "example": "1st Prize"
          },
          "reward_coins": {
            "type": "number",
            "example": 5000
          },
          "reward_gems": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "cz_reward_prize_id",
          "rank",
          "label",
          "reward_coins",
          "reward_gems"
        ]
      },
...
      "RewardLastPlayDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "reward_coins": {
            "type": "number"
          },
          "reward_gems": {
            "type": "number"
          },
          "played_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "label",
          "reward_coins",
          "reward_gems",
          "played_at"
        ]
      },
      "RewardDetailDto": {
        "type": "object",
        "properties": {
          "cz_reward_game_id": {
            "type": "string",
            "format": "uuid"
          },
          "slug": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "instant",
              "draw"
            ]
          },
          "cadence": {
            "type": "string",
            "enum": [
              "none",
              "daily",
              "weekly"
            ]
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "nullable": true
          },
          "headline_prize_coins": {
            "type": "number"
          },
          "entry_cost_gems": {
            "type": "number"
          },
          "status": {
            "type": "string",
            "enum": [
              "live",
              "coming_soon",
              "paused"
            ]
          },
          "is_playable": {
            "type": "boolean"
          },
          "min_entries": {
            "type": "number",
            "example": 1
          },
          "max_entries": {
            "type": "number",
            "example": 250
          },
          "entry_packs": {
            "example": [
              5,
              10,
              25
            ],
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "how_it_works": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardStepDto"
            }
          },
          "terms_url": {
            "type": "string",
            "nullable": true
          },
          "prizes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardPrizeDto"
            }
          },
          "draw": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/LuckyDrawDto"
              }
            ]
          },
          "my_plays_today": {
            "type": "number",
            "nullable": true
          },
          "last_play": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RewardLastPlayDto"
              }
            ]
          },
          "recent_winners": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RewardWinnerDto"
            }
          }
        },
        "required": [
          "cz_reward_game_id",
          "slug",
          "kind",
          "cadence",
          "title",
          "subtitle",
          "icon_url",
          "headline_prize_coins",
          "entry_cost_gems",
          "status",
          "is_playable",
          "min_entries",
          "max_entries",
          "entry_packs",
          "how_it_works",
          "terms_url",
          "prizes",
          "draw",
          "my_plays_today",
          "last_play",
          "recent_winners"
        ]
      },
      "RewardEntriesDto": {
        "type": "object",
        "properties": {
          "cz_lucky_draw_entry_id": {
            "type": "string",
            "format": "uuid"
          },
          "cz_lucky_draw_id": {
            "type": "string",
            "format": "uuid"
          },
          "entries_bought": {
            "type": "number"
          },
          "gems_spent": {
            "type": "number"
          },
          "my_entries": {
            "type": "number"
          },
          "participants_count": {
            "type": "number"
          },
          "entries_count": {
            "type": "number"
          },
          "draw_date": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_lucky_draw_entry_id",
          "cz_lucky_draw_id",
          "entries_bought",
          "gems_spent",
          "my_entries",
          "participants_count",
          "entries_count",
          "draw_date"
        ]
      },
      "BuyEntriesDto": {
        "type": "object",
        "properties": {
          "count": {
            "type": "number",
            "example": 10,
            "description": "How many entries to buy."
          }
        },
        "required": [
          "count"
        ]
      },
      "RewardPlayDto": {
        "type": "object",
        "properties": {
          "cz_reward_play_id": {
            "type": "string",
            "format": "uuid"
          },
          "cz_reward_prize_id": {
            "type": "string",
            "format": "uuid"
          },
          "label": {
            "type": "string"
          },
          "reward_coins": {
            "type": "number"
          },
          "reward_gems": {
            "type": "number"
          },
          "gems_spent": {
            "type": "number"
          },
          "played_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_reward_play_id",
          "cz_reward_prize_id",
          "label",
          "reward_coins",
          "reward_gems",
          "gems_spent",
          "played_at"
        ]
      },
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "email_support",
              "report_problem",
              "feedback"
            ],
            "example": "report_problem"
          },
          "category": {
            "type": "string",
            "example": "Withdrawals",
            "description": "Free text. The app picks it from a list it hard-codes, so options change without a release."
          },
          "issue_type": {
            "type": "string",
            "example": "Coins not credited",
            "description": "Free text, same as category. The app hard-codes the options."
          },
          "occurred_at": {
            "type": "string",
            "example": "2026-08-28T14:30:00.000Z",
            "description": "When the problem happened, from the \"When did the issue happen?\" picker."
          },
          "affected_area": {
            "type": "string",
            "example": "Offers page",
            "description": "Where in the app it happened. Optional on the form."
          },
          "subject": {
            "type": "string",
            "example": "Coins missing after offer"
          },
          "description": {
            "type": "string",
            "example": "I completed the offer yesterday but got no coins."
          },
          "rating": {
            "type": "number",
            "example": 4,
            "description": "Only used for feedback."
          },
          "attachment_urls": {
            "example": [
              "https://cdn.coinzu.app/support/a.png"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "type",
          "description"
        ]
      },
...
      "FaqListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FaqDto"
            }
          },
          "total": {
            "type": "number",
            "example": 12
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FaqCategoryDto"
            }
          }
        },
        "required": [
          "data",
          "total",
          "categories"
        ]
      },
...
    }
  }
}